requirements.txt
termcolor==1.1.0
colorama==0.3.9
pyfiglet==0.7.5
pycrypto==2.6.1
pyserial==3.4
